--- specs/src/main/java/org/seedstack/business/api/Creatable.java
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) 2013-2015 by The SeedStack authors. All rights reserved.
- *
- * This file is part of SeedStack, An enterprise-oriented full development stack.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package org.seedstack.business.api;
-
-/**
- * Interface that marks objects implementing it as being able to be created.
- * 
- * @author epo.jemba@ext.mpsa.com
- * 
- */
-public  interface Creatable {}
