+++ specs/src/main/java/org/seedstack/business/spi/GenericImplementation.java
@@ -7,7 +7,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.business.api;
+package org.seedstack.business.spi;
 
 import java.lang.annotation.*;
 
@@ -22,5 +22,5 @@ import java.lang.annotation.*;
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
-public @interface DefaultImpl {
+public @interface GenericImplementation {
 }
